Implement notification support in `SimpleSwitch` (#1167)
This PR implements support for notifications (idle timeouts, digests) in `SimpleSwitch` when build with PI.

The code of `NotificationsCapture` is taken verbatim from `SimpleSwitchGrpc`.

Context: Stratum is consuming the `SimpleSwitch[Runner]` downstream without gRPC support, as we're implementing that ourselves on top. We'd still like to use digests and idle timeouts, which can be delivered through the PI layer.