Fix simple_switch CLI commands to configure queues (#1276)
The queue id for the `set_queue_rate` and `set_queue_depth` commands
did not match the "datapath" queue id, as used in the P4 program.

This patch fixes the inconsistency by changing the priority queue implementation
so that the queue with the highest numerical priority is served first. For the P4
programmer, the observable behavior is not changed: queue with priority 0 is
the lowest priority queue. However, the CLI commands now work as expected.

Fixes #1264 

Signed-off-by: Rizky Ramadhana <rizkyramadhana26@gmail.com>