Adding pcap_mock test; finding and fixing bugs using it.
Added a libpcap mock for testing the BMI code in P4 behavioral-model by injecting packets
using the actual BMI code instead of their nanomsg based injector (which is bypassing
the BMI code)
It is not meant to be a full mock of libpcap, it only implements the functions required
for this special purpose.

Bugs fixed

bmi_port.c:

In bmi_port_destroy_mgr function,
remove interfaces in reverse order. The previous code removed only half of the interfaces
because it was iterating through an array while removing its members.

In _bmi_port_interface_remove function,
bug fix to prevent a deadlock. Function uses "write" to write on a pipe and waits for the
answer using "read" while it holds the lock "port_mgr->lock". Meanwhile, the "run_select"
thread requires to get that lock once per iteration of the loop, causing a deadlock.

dev_mgr_bmi.cpp:

In BmiDevMgrImp class,
p_monitor should be stopped in destructor and before destruction of port_mgr to prevent
pure virtual function call.